requirements.txt
fastapi
uvicorn
pillow
tensorflow
numpy
redis
python-dotenv
python-multipart
